requirements.txt
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
open_clip_torch==3.2.0
packaging==25.0
pillow==12.1.0
PyYAML==6.0.3
regex==2025.11.3
safetensors==0.7.0
setuptools==70.2.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.24
torch==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
wcwidth==0.2.14
